server/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core
langchain-groq

pinecone
langchain-huggingface
sentence-transformers
torch

PyPDF

python-dotenv

pydantic
requests
tqdm

loguru
streamlit
